+++ android/library/core/src/main/java/com/liferay/mobile/screens/audiencetargeting/interactor/loadscreenlets/AudienceTargetingLoadScreenletsInteractorImpl.java
@@ -1,18 +1,11 @@
 package com.liferay.mobile.screens.audiencetargeting.interactor.loadscreenlets;
 
 import com.liferay.mobile.android.service.Session;
-import com.liferay.mobile.android.v62.dlfileentry.DLFileEntryService;
 import com.liferay.mobile.screens.audiencetargeting.AudienceTargetingListener;
 import com.liferay.mobile.screens.audiencetargeting.UserContext;
-import com.liferay.mobile.screens.audiencetargeting.interactor.AudienceTargetingCallback;
-import com.liferay.mobile.screens.audiencetargeting.interactor.AudienceTargetingResult;
-import com.liferay.mobile.screens.audiencetargeting.interactor.requestcontent.AudienceTargetingContentEvent;
 import com.liferay.mobile.screens.base.interactor.BaseRemoteInteractor;
-import com.liferay.mobile.screens.base.interactor.JSONObjectCallback;
-import com.liferay.mobile.screens.base.interactor.JSONObjectEvent;
 import com.liferay.mobile.screens.context.SessionContext;
 import com.liferay.mobile.screens.service.v62.ScreensAudienceTargetingService;
-import com.liferay.mobile.screens.util.LiferayLogger;
 
 import org.json.JSONObject;
 
@@ -27,13 +20,13 @@ public class AudienceTargetingLoadScreenletsInteractorImpl
 		super(targetScreenletId);
 	}
 
-	public void getScreenlets(String screenletApp, Integer groupId, UserContext userContext)
+	public void getScreenlets(String screenletApp, Long groupId, UserContext userContext)
 		throws Exception {
 
 		getScreenlets(screenletApp, groupId, null, userContext);
 	}
 
-	public void getScreenlets(String screenletApp, Integer groupId, String placeholder, UserContext userContext)
+	public void getScreenlets(String screenletApp, Long groupId, String placeholder, UserContext userContext)
 		throws Exception {
 
 		JSONObject jsonObject = userContext.toJSON();
